Core refactoring (not tested)
