+++ acalib/algorithms/hrep.py
@@ -1,5 +1,38 @@
 import acalib
 from .algorithm import Algorithm
+from astropy.nddata import support_nddata
+
+# TODO: This is non-generic. Try to use the UPI (it can be done!)
+@support_nddata
+def vel_stacking(data,data_slice,wcs=None,uncertainty=None, mask=None, meta=None, unit=None):
+     """
+     Create an image collapsing the frecuency axis
+
+     Parameters
+     ----------
+     data : numpy.ndarray or astropy.nddata.NDData or astropy.nddata.NDDataRef
+         Astronomical 2D image
+
+     slice : slice object
+         Sector to be collapsed
+
+     Returns
+     -------
+     image (NDDataRef): 2D-Array with the stacked cube.
+
+     """
+     if len(data.shape) != 3:
+         log.error("Cube needs to be a 3D array")
+         raise ValueError("Cube needs to be a 3D array")
+     dims = data.shape
+     subcube = data[data_slice, :,:]
+     stacked = np.sum(subcube,axis=0)
+     if wcs:
+         wcs = wcs.dropaxis(2)
+
+         return NDDataRef(stacked, uncertainty=uncertainty, mask=mask,wcs=wcs, meta=meta, unit=unit)
+     else:
+         return stacked
 
 
 class ScatterPixelRepresentation(Algorithm):
@@ -27,7 +60,7 @@ class ScatterPixelRepresentation(Algorithm):
 
         pp_slices = []
         for slice in slices:
-            pp_slice = acalib.core.vel_stacking(data.data, slice)
+            pp_slice = vel_stacking(data.data, slice)
             labeled_images = acalib.core.gaussian_mix(pp_slice, prob=self.config["P"],
                                                       precision=self.config["PRECISION"])
