+++ acalib/core/__init__.py
@@ -1,5 +1,5 @@
-from .analysis import *
-from .models import *
+from .image_analysis import *
+from .function_models import *
 from .utils import *
 from .transform import *
-from .compact import *
+from .compact_rep import *
