+++ acalib/core/compact_rep.py
@@ -2,7 +2,7 @@ from astropy import log
 from astropy.table import Table
 import numpy as np
 from . import matching_slabs, fix_limits, slab, snr_estimation, create_mould, add
-from .models import _eighth_mould
+from .function_models import _eighth_mould
 
 
 def _update_min_energy(energy,mat,ub,lb,delta):
@@ -198,144 +198,3 @@ def precision_from_delta(delta,clev):
     P=np.diag(sq_delta)
     return(-2*np.log(clev)*P)
 
-
-#def gmr_from_mould(data,threshold,nlevel,P,upper=None,lower=None,max_iter=None,full_output=False):
-#    debug=True
-#    inten=[]
-#    center=[]
-#    struct=[]
-#    niter=0
-#    message="Not finished"
-#
-#    #Restrict data to what the corresponding slab
-#    data=data[slab(data,upper,lower)]
-#    if max_iter==None:
-#        d=data.ndim
-#        #GMRs are useless if we have more data than the original data
-#        max_iter=data.size/(1 + d + d*d)
-#    residual=data.copy()
-#    # Argmax
-#    max_idx=np.unravel_index(residual.argmax(),residual.shape)
-#    max_val=residual[max_idx]
-#    # Compute delta
-#    Sigma=np.linalg.inv(P)
-#    delta=np.round(np.sqrt(2*np.log(max_val/nlevel)*Sigma.diagonal()))
-#    # Compute mould TODO
-#    mould=create_mould(P,delta)
-#    #plt.imshow(mould[0,:,:])
-#    #plt.colorbar()
-#    #plt.show()
-#    #plt.imshow(mould[1,:,:])
-#    #plt.colorbar()
-#    #plt.show()
-#    #plt.imshow(mould[2,:,:])
-#    #plt.colorbar()
-#    #plt.show()
-#    #plt.imshow(mould[3,:,:])
-#    #plt.colorbar()
-#    #plt.show()
-#    #plt.imshow(mould[4,:,:])
-#    #plt.colorbar()
-#    #plt.show()
-#    #mould=mould/mould.max()
-#    #discretize delta
-#    delta=np.floor(np.array(mould.shape)/2)
-#    # Create energy matrix
-#    energy=np.zeros(data.shape)
-#    #np.empty_like(data)
-#    #mask=np.isnan(energy)
-#    #energy.data[np.logical_not(mask)]=max_val
-#    # check for diagonality
-#    #dstruct=np.all(a == np.diag(np.diag(P)))
-#    # compute 
-#    if debug:
-#       log.info("Delta = "+str(delta))
-#       log.info("Max Iter = "+str(max_iter))
-#       #plt.imshow(mould.sum(axis=(0)))
-#       #plt.show()
-#    _update_energies(energy,residual,mould,nlevel,delta)
-#    
-#    while True:
-#       niter+=1
-#       if (niter > max_iter):
-#           message="Maximum iterations reached = "+str(niter)
-#           break
-#       max_idx=np.array(np.unravel_index(energy.argmax(),energy.shape))
-#       max_val=energy[tuple(max_idx)]
-#       #    tlb=base-delta
-#       #    tub=base+delta + 1
-#       #    plt.imshow(residual[slab(residual,tlb,tub)].sum(axis=0))
-#       #    print energy[tuple(base)]
-#       #    print residual[slab(residual,tlb,tub)]
-#       #    print mould
-#       #    print (residual[slab(residual,tlb,tub)]/mould)
-#       #    plt.show()
-#       rem=max_val - nlevel
-#       
-#       if rem <= 0.0:
-#           message="No more signal to extract at iteration "+str(niter)
-#           break
-#       if (rem < threshold):
-#           message="Signal reached the desired threshold of "+str(threshold)+" at iteration "+str(niter)
-#           break
-#       inten.append(rem)
-#       center.append(max_idx)
-#       struct.append(P)
-#       lb=max_idx - delta
-#       ub=max_idx + delta +1 
-#       add_flux(residual,-rem*mould,lb,ub)
-#       _update_energies(energy,residual,mould,nlevel,delta,lower=lb-delta,upper=ub+delta)
-#       if debug:
-#           log.info("Iter "+str(niter))
-#           log.info("Maximum energy E = "+str(max_val)+" at "+str(max_idx))
-#           log.info("Remove E = "+str(rem)+" SNR = "+str(rem/nlevel))# + " GAP = "+ str(y/rms - 1.0 - snrlimit))
-#           print "max_energy = ",np.max(energy)
-#           print "min_energy = ",np.min(energy)
-#           print "max_resid = ",np.max(residual)
-#           print "min_resid = ",np.min(residual)
-#           base=np.unravel_index(residual.argmax(),residual.shape)
-#           print "e(max_resid)=",energy[tuple(base)]
-#           #plt.imshow(residual.sum(axis=(0)))
-#           #plt.colorbar()
-#           #plt.show()
-#           #plt.imshow(energy.sum(axis=(0)))
-#           #plt.colorbar()
-#           #plt.show()
-#    inten=np.array(inten)
-#    center=np.array(center)
-#    struct=np.array(struct)
-#    res=Table([inten,center,struct],names=('intensity','center','struct'))
-#    if full_output:
-#       return res,message,residual,energy
-#    return res
-# 
-# 
-#
-#if __name__ == '__main__':
-#    # SandBox space for testing
-#    a=3*np.random.random((50,200,200))
-#    P=np.array([[2,0,0],[0,1,0],[0,0,1]])
-#    freak=np.array([[2,0.5,0.9],[0.5,1,0.3],[0.9,0.3,1]])
-#    delta=np.array([10,50,50])
-#    p1=np.array([15,97,143])
-#    peak=create_mould(0.01*freak,delta)
-#    add_flux(a,5*peak,p1-delta,p1+delta+1)
-#    p2=np.array([30,45,97])
-#    peak=create_mould(0.02*freak,delta)
-#    add_flux(a,10*peak,p2-delta,p2+delta+1)
-#    rms=estimate_rms(a)
-#    print("RMS = "+str(rms))
-#    plt.imshow(np.sum(a,axis=(0)))
-#    plt.colorbar()
-#    plt.show()
-#    (tab,message,residual,energy)=gmr_from_mould(a,0.2*rms,rms,P,full_output=True)
-#    print(message)
-#    print(tab)
-#    plt.imshow(np.sum(residual,axis=(0)))
-#    plt.colorbar()
-#    plt.show()
-#    plt.imshow(np.sum(energy,axis=(0)))
-#    plt.colorbar()
-#    plt.show()
-#
-#
